+++ build.sbt
@@ -3,7 +3,7 @@ ThisBuild / name := "cards"
 ThisBuild / organization := "io.github.jmisabella"
 ThisBuild / organizationName := "jmisabella"
 ThisBuild / organizationHomepage := Some(url("https://github.com/jmisabella"))
-ThisBuild / version := "0.1.11"
+ThisBuild / version := "0.1.12"
 ThisBuild / scalaVersion := "2.13.10"
 
 ThisBuild / scmInfo := Some(
