+++ src/test/scala/behaviors/play/BlackjackPlaySpec.scala
@@ -992,74 +992,4 @@ class BlackjackPlaySpec extends AnyFlatSpec with GivenWhenThen {
     nextHandIndex should equal (1)
   }
 
-  it should "not allow player or dealer to continue playing after player with 3 card hand has busted" in {
-    Given("a game with 1 player having 3 cards and whose hand is busted, a dealer with no cards")
-    val player1 = BlackjackPlayerState(
-      "Jeffrey", 
-      25, 
-      Seq( 
-        Hand(hand = Seq(Card(Ten, Hearts), Card(Eight, Clubs), Card(Four, Clubs)), bets = Map("Jeffrey" -> 5), outcome = None)) )
-    val history = Seq(Action("Jeffrey", Hit, Seq(Card(Four, Clubs)), None, Seq(Card(Ten, Hearts), Card(Eight, Clubs)), Seq(Seq(Card(Ten, Hearts), Card(Eight, Clubs), Card(Four, Clubs)))))
-    val game = BlackjackGameState(
-      options = BlackjackOptions(dealerHitLimit = H17), 
-      minimumBet = 5, 
-      players = Seq(player1), 
-      history = history, 
-      currentPlayerIndex = Some(0),
-      currentHandIndex = Some(0))
-    When("performing Hit play action")
-    val (updatedHands, updatedDeck, updatedHistory) = performPlayAction("Jeffrey", Hit, game.currentHand(), game.deck)
-    Then("round should end with player Busting before Losing")
-    updatedHistory.reverse.head.action should equal (Lose)
-    updatedHistory.reverse.head.playerId should equal ("Jeffrey")
-    updatedHistory.reverse.tail.head.action should equal (Bust)
-    updatedHistory.reverse.tail.head.playerId should equal ("Jeffrey")
-  }
-
-  it should "not allow player or dealer to continue playing after player with 3 card hand has achieved 21" in {
-    Given("a game with 1 player having 3 cards and whose hand equals 21, a dealer with no cards")
-    val hand = Seq(Card(Ten, Hearts), Card(Eight, Clubs), Card(Three, Clubs))
-    val player1 = BlackjackPlayerState(
-      "Jeffrey", 
-      25, 
-      Seq( 
-        Hand(hand = hand, bets = Map("Jeffrey" -> 5), outcome = None)) )
-    val history = Seq(Action("Jeffrey", Hit, Seq(Card(Four, Clubs)), None, hand, Seq(hand)))
-    val game = BlackjackGameState(
-      options = BlackjackOptions(dealerHitLimit = H17), 
-      minimumBet = 5, 
-      players = Seq(player1), 
-      history = history, 
-      currentPlayerIndex = Some(0),
-      currentHandIndex = Some(0))
-    When("performing Hit play action")
-    val (updatedHands, updatedDeck, updatedHistory) = performPlayAction("Jeffrey", Hit, game.currentHand(), game.deck)
-    Then("round should end with player Busting before Losing")
-    updatedHistory.reverse.head.action should equal (Win)
-    updatedHistory.reverse.head.playerId should equal ("Jeffrey")
-  }
-
-  it should "not allow player or dealer to continue playing after player has been dealt a blackjack" in {
-    Given("a game with 1 player having 2 cards and whose hand equals 21, a dealer with no cards")
-    val hand = Seq(Card(Ten, Hearts), Card(Ace, Clubs))
-    val player1 = BlackjackPlayerState(
-      "Jeffrey", 
-      25, 
-      Seq( 
-        Hand(hand = hand, bets = Map("Jeffrey" -> 5), outcome = None)) )
-    val game = BlackjackGameState(
-      options = BlackjackOptions(dealerHitLimit = H17), 
-      minimumBet = 5, 
-      players = Seq(player1),
-      currentPlayerIndex = Some(0),
-      currentHandIndex = Some(0))
-    When("playing player's hand")
-    var result = playHand(game)
-    Then("player should have actions Blackjack and Win recorded in history")
-    result.history.reverse.head.playerId should equal ("Jeffrey")
-    result.history.reverse.head.action should equal (Win)
-    result.history.reverse.tail.head.playerId should equal ("Jeffrey")
-    result.history.reverse.tail.head.action should equal (Blackjack)
-  }
-
 }
